[Gold V] Title: 하노이 탑 이동 순서, Time: 544 ms, Memory: 40648 KB -BaekjoonHub

+++ 백준/Gold/11729. 하노이 탑 이동 순서/하노이 탑 이동 순서.java
@@ -4,7 +4,7 @@ public class Main{
 
     public static void main(String[] args){
         Scanner scan = new Scanner(System.in);
-        int n = scan.nextByte();
+        int n = scan.nextInt();
 
         System.out.println((1<<n) -1);
         StringBuilder sb = new StringBuilder();
